pwa-v2/backend-requirements.txt
# RelatoRecibo Backend - Python Dependencies
# Production requirements

# ===== FastAPI Core =====
fastapi==0.104.1
uvicorn[standard]==0.24.0  # ASGI server with websocket support
pydantic==2.5.0            # Data validation
pydantic-settings==2.1.0   # Settings management

# ===== Database & Supabase =====
supabase==2.3.0            # Supabase Python client
postgrest==0.13.0          # PostgreSQL REST client
httpx==0.25.2              # Async HTTP client (required by supabase)

# ===== Authentication & Security =====
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4            # Password hashing
python-multipart==0.0.6           # Form data & file uploads

# ===== Image Processing & OCR =====
Pillow==10.1.0             # Image processing
pytesseract==0.3.10        # Tesseract OCR wrapper
numpy==1.26.2              # Array processing for image operations

# ===== PDF Generation =====
reportlab==4.0.7           # PDF generation
pypdf==3.17.1              # PDF manipulation (optional)

# ===== Utilities =====
python-dotenv==1.0.0       # Environment variables
loguru==0.7.2              # Logging
python-dateutil==2.8.2     # Date utilities

# ===== Performance & Caching =====
aiofiles==23.2.1           # Async file operations
cachetools==5.3.2          # Caching utilities

# ===== Rate Limiting =====
slowapi==0.1.9             # Rate limiting for FastAPI

# ===== Monitoring (Optional) =====
# sentry-sdk==1.38.0       # Error tracking (uncomment if needed)
